[Refactor] PageController 설정 및 로그인 후 네브바 수정
