+++ src/main/java/com/mycom/myapp/naviya/global/security/config/SecurityConfig.java
@@ -20,8 +20,8 @@ public class SecurityConfig {
 
         http
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/", "/login", "/signup", "/signupProc", "/userpage", "/userpage/update", "/main"
-,                                "/static/**", "/css/**", "/js/**", "/images/**").permitAll()
+                        .requestMatchers("/", "/login", "/signup", "/signupProc",
+                                        "/static/**", "/css/**", "/js/**", "/images/**").permitAll()
                         .anyRequest().authenticated()
                 )
                 .formLogin(auth -> auth
@@ -34,7 +34,7 @@ public class SecurityConfig {
                 )
                 .logout(logout -> logout
                         .logoutUrl("/logout")
-                        .logoutSuccessUrl("/login")
+                        .logoutSuccessUrl("/")
                 )
                 .csrf(csrf -> csrf.disable());
