+++ src/main/resources/templates/main.html
@@ -1,15 +1,15 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Main Page</title>
-</head>
-<body>
-<h1>MAIN페이지입니당</h1>
-<hr>
-<h1>Welcome, <span th:text="${email}">User</span>!</h1>
+<!--<!DOCTYPE html>-->
+<!--<html xmlns:th="http://www.thymeleaf.org">-->
+<!--<head>-->
+<!--    <title>Main Page</title>-->
+<!--</head>-->
+<!--<body>-->
+<!--<h1>MAIN페이지입니당</h1>-->
+<!--<hr>-->
+<!--<h1>Welcome, <span th:text="${email}">User</span>!</h1>-->
 
-<p>EMAIL: <span th:text="${email}">email</span></p>
-<p>Role: <span th:text="${role}">Role</span></p>
+<!--<p>EMAIL: <span th:text="${email}">email</span></p>-->
+<!--<p>Role: <span th:text="${role}">Role</span></p>-->
 
-</body>
-</html>
+<!--</body>-->
+<!--</html>-->
