+++ src/main/resources/templates/userpage.html
@@ -43,9 +43,131 @@
         .form-check-input {
             margin-left: 10px;
         }
+
+
+        body {
+            overflow-x: hidden;
+        }
+
+        /* Navigation Bar styles */
+        .navbar {
+            background-color: #47688d; /* 밝은 색으로 변경 (흰색) */
+            border-bottom: 1px solid #e0e0e0; /* 아래쪽에 얇은 테두리 추가 */
+        }
+
+        .navbar-nav .nav-link {
+            color: white !important; /* 어두운 색으로 텍스트 변경 */
+        }
+
+        .navbar-nav .nav-link:hover {
+            color: #c4d6e9 !important; /* 파란색으로 변경 (hover 시) */
+        }
+
+        /* Sidebar styles */
+        #sidebar {
+            height: 100%;
+            width: 250px;
+            position: fixed;
+            top: 0;
+            left: -250px; /* 사이드바는 기본적으로 숨김 */
+            background-color: #f7f7f7; /* 연한 회색으로 변경 */
+            color: black;
+            transition: left 0.3s ease; /* 부드러운 전환 효과 */
+            padding-top: 20px;
+            z-index: 1000;
+        }
+
+        #sidebar a {
+            padding: 15px;
+            text-decoration: none;
+            font-size: 18px;
+            color: black;
+            display: block;
+            transition: background-color 0.2s;
+        }
+
+        #sidebar a:hover {
+            background-color: #e0e0e0; /* hover 시 배경색을 밝은 회색으로 변경 */
+        }
+
+        #sidebar a:first-child {
+            margin-top: 35px; /* 필요에 따라 조정 */
+        }
+
+        /* Toggle Button styles */
+        #toggleSidebar {
+            font-size: 24px;
+            cursor: pointer;
+            position: fixed;
+            top: 15px;
+            left: 15px;
+            z-index: 1100; /* 다른 요소 위에 표시 */
+            transition: left 0.3s ease; /* 사이드바에 따라 버튼 이동 */
+            /* background-color: #343a40; 어두운 색으로 설정 */
+            color: white; /* 흰색 텍스트 */
+            padding: 10px; /* 클릭 영역을 위한 패딩 */
+            border-radius: 5px; /* 둥근 모서리 */
+        }
+
+        /* 사이드바 활성 상태 */
+        .sidebar-active {
+            left: 0; /* 사이드바를 표시 */
+        }
+
+        /* 사이드바가 활성 상태일 때 토글 버튼 이동 */
+        .toggle-active {
+            left: 265px; /* 사이드바와 정렬 */
+        }
+
+        #main-content {
+            transition: margin-left 0.3s ease;
+        }
     </style>
 </head>
 <body>
+
+
+<!-- Navigation Bar -->
+<nav class="navbar navbar-expand-lg navbar">
+    <div class="container-fluid">
+        <!-- Toggle Button -->
+        <span id="toggleSidebar" class="navbar-toggler-icon"></span>
+
+        <!-- Right-Side Links -->
+        <div class="collapse navbar-collapse justify-content-end">
+            <ul class="navbar-nav">
+                <li class="nav-item">
+                    <a class="nav-link" href="/">메인페이지</a>
+                </li>
+                <!-- 로그인 상태 확인 -->
+                <li class="nav-item" th:if="${user == null}">
+                    <a class="nav-link" href="/login">로그인</a>
+                </li>
+                <li class="nav-item" th:if="${user == null}">
+                    <a class="nav-link" href="/signup">회원가입</a>
+                </li>
+                <!-- 로그인한 사용자 정보 -->
+                <li class="nav-item" th:if="${user != null}">
+                    <a class="nav-link" th:text="${user.name + '님'}"></a>
+                </li>
+                <li class="nav-item" th:if="${user != null}">
+                    <a class="nav-link" href="/logout">로그아웃</a>
+                </li>
+            </ul>
+        </div>
+    </div>
+</nav>
+
+<!-- Sidebar -->
+<div id="sidebar">
+    <a href="#">아이 프로필 변경</a>
+    <a href="#">아이정보 보기</a>
+    <a href="/userpage">계정보기</a>
+    <a href="#">최근 책 보기</a>
+    <a href="/ChildFavorBookList.html">좋아요 누른 책 보기</a>
+    <a href="#">카테고리 별 책보기</a>
+    <a href="/lottery/form">선착순 이벤트</a>
+</div>
 
 
 <div class="container">
@@ -141,6 +263,24 @@
                 console.error('Error updating user data:', error);
                 alert("회원정보 수정에 실패했습니다.");
             });
+    });
+</script>
+
+<!-- JavaScript for sidebar toggle -->
+<script>
+    // 사이드바 토글 기능
+    document.getElementById('toggleSidebar').addEventListener('click', function() {
+        const sidebar = document.getElementById('sidebar');
+        const toggleBtn = document.getElementById('toggleSidebar');
+
+        // 사이드바 가시성 전환
+        if (sidebar.style.left === '0px') {
+            sidebar.style.left = '-250px'; // 사이드바 숨기기
+            toggleBtn.classList.remove('toggle-active'); // 버튼 원래대로
+        } else {
+            sidebar.style.left = '0px'; // 사이드바 보이기
+            toggleBtn.classList.add('toggle-active'); // 버튼 함께 이동
+        }
     });
 </script>
